Moved Plugin type to schemas package
