+++ packages/gb-printer-schemas/src/index.ts
@@ -3,5 +3,6 @@ export * from '@/schemas/items/FrameGroup';
 export * from '@/schemas/items/Image';
 export * from '@/schemas/items/ImageGroup';
 export * from '@/schemas/items/Palette';
+export * from '@/schemas/items/Plugin';
 export * from '@/tools/creationDate';
 export * from '@/tools/safeDate';
