+++ packages/gb-printer-web/src/consts/plugins.ts
@@ -5,10 +5,3 @@ export const CompatibilityActionType = {
   IMPORT_FILES: 'IMPORT_FILES',
 } as const;
 export type CompatibilityActionType = (typeof CompatibilityActionType)[keyof typeof CompatibilityActionType];
-
-export const ConfigParamType = {
-  NUMBER: 'number',
-  STRING: 'string',
-  MULTILINE: 'multiline',
-} as const;
-export type ConfigParamType = (typeof ConfigParamType)[keyof typeof ConfigParamType];
