+++ packages/gb-printer-web/src/types/Plugin.ts
@@ -1,8 +1,6 @@
 import type FileSaver from 'file-saver';
 import { type RGBNTiles, type RGBNPalette, type ExportFrameMode } from 'gb-image-decoder';
-import { type Palette , type Image } from 'gb-printer-schemas';
-import z from 'zod';
-import { ConfigParamType } from '@/consts/plugins';
+import { type Palette, type Image, type PluginConfigValues, type PluginConfigParams } from 'gb-printer-schemas';
 import { type ImportFn } from '@/hooks/useImportExportSettings';
 import { UsePlugins } from '@/hooks/usePlugins';
 import { type UseStores } from '@/hooks/useStores';
@@ -21,37 +19,6 @@ export interface PluginFunctions {
   addImages: (images: Image[]) => void;
   alert: (title: string, text: string) => void;
 }
-
-export const ConfigParamSchema = z.object({
-  label: z.string(),
-  type: z.enum(ConfigParamType),
-});
-
-export type ConfigParam = z.infer<typeof ConfigParamSchema>;
-
-export const PluginConfigParamsSchema = z.partialRecord(
-  z.string(),
-  ConfigParamSchema,
-).default({});
-export type PluginConfigParams = z.infer<typeof PluginConfigParamsSchema>;
-
-export const PluginConfigValuesSchema = z.partialRecord(
-  z.string(),
-  z.union([z.number(), z.string()],
-)).default({});
-export type PluginConfigValues = z.infer<typeof PluginConfigValuesSchema>;
-
-export const PluginSchema = z.object({
-  url: z.string(),
-  config: PluginConfigValuesSchema.optional(),
-  name: z.string().prefault(''),
-  description: z.string().prefault(''),
-  // loading: z.boolean().optional(),
-  // error: z.union([z.string(), z.literal(false)]).optional(),
-  configParams: PluginConfigParamsSchema.optional(),
-});
-
-export type Plugin = z.infer<typeof PluginSchema>;
 
 export interface GetCanvasOptions {
   scaleFactor?: number;
