+++ packages/gb-printer-web/src/workers/itemsIndexedDbWorker/schemas.ts
@@ -3,10 +3,10 @@ import { type Image, ImageSchema } from 'gb-printer-schemas';
 import { SerializableImageGroupSchema } from 'gb-printer-schemas';
 import { FrameSchema } from 'gb-printer-schemas';
 import { FrameGroupSchema } from 'gb-printer-schemas';
+import { PluginSchema } from 'gb-printer-schemas';
 import z from 'zod';
 import { SpecialTags } from '@/consts/SpecialTags';
 import { BinaryStoreItemSchema } from '@/types/BinaryStoreItem';
-import { PluginSchema } from '@/types/Plugin';
 
 const DAY_MS = 24 * 60 * 60 * 1000;
